requirements.txt
# Core MQTT Client
paho-mqtt==1.6.1

# Modbus RTU/RS485 Support
pymodbus==3.5.4
pyserial==3.5

# I2C Support
smbus2==0.4.3

# Configuration Management
pyyaml==6.0.1

# Data Validation
pydantic==2.5.3

# Logging and Monitoring
python-json-logger==2.0.7

# Retry and Backoff Logic
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# Development Tools
black==23.12.1
flake8==6.1.0
mypy==1.7.1
